backend/requirements.txt
torch==2.0.1
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
Pillow==9.5.0
fastapi==0.104.1
redis-om==0.2.1
uvicorn==0.24.0.post1